% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arc_footer.R
\name{add_arc_footer}
\alias{add_arc_footer}
\title{Add The Arc logo and URL footer to a ggplot2 plot}
\usage{
add_arc_footer(plot = ggplot2::last_plot())
}
\arguments{
\item{plot}{ggplot2 chart to which a footer should be appended. Default is
`ggplot2::last_plot()`}
}
\description{
Add The Arc logo and URL footer to a ggplot2 plot
}
