% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_long.R
\name{add_team_long}
\alias{add_team_long}
\title{Add a column containing long team names to a dataframe with a `team` column}
\usage{
add_team_long(df)
}
\arguments{
\item{df}{A dataframe containing a column called `team` with short
(two-character) team names, as in `arcfooty::team_names`}
}
\description{
Add a column containing long team names to a dataframe with a `team` column
}
