% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_save.R
\name{arc_save}
\alias{arc_save}
\alias{arc_save_tall}
\alias{arc_save_very_tall}
\alias{arc_save_extremely_tall}
\alias{arc.save}
\alias{arc.save.tall}
\alias{arc.save.very.tall}
\alias{arc.save.extremely.tall}
\title{Save ggplot2 plots in the right size}
\usage{
arc_save(
  filename,
  object = ggplot2::last_plot(),
  width = 169.333,
  type = "normal",
  height = NULL
)

arc_save_tall(filename, object = ggplot2::last_plot())

arc_save_very_tall(filename, object = ggplot2::last_plot())

arc_save_extremely_tall(filename, object = ggplot2::last_plot())

arc.save(
  filename,
  object = ggplot2::last_plot(),
  width = 169.333,
  type = "normal",
  height = NULL
)

arc.save.tall(filename, object = ggplot2::last_plot())

arc.save.very.tall(filename, object = ggplot2::last_plot())

arc.save.extremely.tall(filename, object = ggplot2::last_plot())
}
\arguments{
\item{filename}{Name of the file the save, including extension.}

\item{object}{ggplot2 object to save.}

\item{width}{169.333}

\item{type}{Chart type (ie. height). Options include:
\itemize{ \item{"normal"}{ The default. 103.18756mm}
          \item{"tall"}{ 127mm}
          \item{"very_tall"}{ 169.333mm}
          \item{"extremely_tall"}{ 206.375mm}
}}

\item{height}{NULL by default. When non-null,
overrides the value set by 'type'.}
}
\description{
Save ggplot2 plots in the right size
}
