% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{team_cols}
\alias{team_cols}
\title{AFL team colours}
\format{named numeric vector}
\usage{
team_cols
}
\description{
A character vector containing hex codes corresponding to the team colours
for AFL/VFL teams. The vector is named.
}
\keyword{datasets}
