% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{team_names}
\alias{team_names}
\title{AFL team names in various forms}
\format{tibble}
\usage{
team_names
}
\description{
A tibble containing AFL/VFL team names in four formats: short
(two character) name, `team`, longer name as used at The Arc (`team_long`),
and the short (`espn_team`) and long (`espn_team_long`) versions used
by ESPN.
}
\keyword{datasets}
