% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_arc.R
\name{theme_arc}
\alias{theme_arc}
\title{Style ggplot2 charts in the The Arc way}
\usage{
theme_arc(
  base_size = 12,
  base_family = "Lato",
  base_line_size = base_size/22,
  base_rect_size = 0
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family; default is "Lato". A warning will be
shown if Lato not available to R.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements; default is 0}
}
\description{
Style ggplot2 charts in the The Arc way
}
\examples{

library(ggplot2)

\donttest{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
      geom_point() +
      theme_arc()

p
}

}
